--- src/integration_tests/data/dataset_simple.json
@@ -1,15 +1,0 @@
-{
-  "title": "SPPI supplementary data set",
-  "description": "supplementary data for SPPI Survey",
-  "schema_version" : 1,
-  "survey_id": "Survey 1",
-  "ce_id": "2312",
-  "data": [
-    {
-      "unit_id": "55e64129-6acd-438b-a23a-3cf9524ab912"
-    },
-    {
-      "unit_id": "f6f1c355-5465-4063-aa7a-f290e649aa00"
-    }
-  ]
-}
