--- src/integration_tests/data/publish_dataset.json
@@ -1,3 +1,0 @@
-{
-  "dataset_id": "2bd66ae3-e66d-4c02-b83a-1bf518e42edb"
-}
