--- src/integration_tests/data/query_datasets_by_period_id.json
@@ -1,14 +1,0 @@
-{
-  "period_id": "2312",
-  "datasets": [
-    {
-      "dataset_id": "3518d94f-7273-4b89-9d1a-eb4bcea98a80",
-      "title": "SPPI supplementary data set",
-      "description": "supplementary data for SPPI Survey",
-      "schema_version": 1,
-      "survey_id": "abc",
-      "survey_type": "business",
-      "time_published": "2023-01-12T16:37:06Z"
-    }
-  ]
-}
