--- src/integration_tests/data/query_datasets_by_survey_id.json
@@ -1,32 +1,0 @@
-{
-  "survey_id": "abc",
-  "datasets": [
-    {
-      "dataset_id": "3518d94f-7273-4b89-9d1a-eb4bcea98a80",
-      "title": "SPPI supplementary data set",
-      "description": "supplementary data for SPPI Survey",
-      "schema_version": 1,
-      "period_id": "2312",
-      "survey_type": "business",
-      "time_published": "2023-01-12T16:37:06Z"
-    },
-    {
-      "dataset_id": "35141b5a-7afc-4a6c-857f-1a94cc9d4371",
-      "title": "SPPI supplementary data set",
-      "description": "supplementary data for SPPI Survey",
-      "schema_version": 1,
-      "period_id": "2313",
-      "survey_type": "business",
-      "time_published": "2023-01-13T16:37:06Z"
-    },
-    {
-      "dataset_id": "eafbbae8-37cb-44e0-998e-c7b73f6385cc",
-      "title": "I changed the title",
-      "description": "and the description",
-      "schema_version": 2,
-      "period_id": "2314",
-      "survey_type": "business",
-      "time_published": "2023-01-14T16:37:06Z"
-    }
-  ]
-}
