--- src/integration_tests/data/query_schemas.json
@@ -1,5 +1,0 @@
-{
-  "survey_id": "abc",
-  "latest_version": 3,
-  "versions": [1,2,3]
-}
