--- src/integration_tests/data/retrieve_unit_data.json
@@ -1,23 +1,0 @@
-{
-  "unit_id": "55e64129-6acd-438b-a23a-3cf9524ab912",
-  "properties": {
-    "sample_unit": {
-      "units_of_sale": "MILES MAPPED",
-      "currency_description": "SILVER COINS",
-      "time_items": [
-        {
-          "ref": "M1",
-          "grade": "Chief mapper"
-        },
-        {
-          "ref": "M2",
-          "grade": "Junior mapper"
-        },
-        {
-          "ref": "M3",
-          "grade": "Bag carrier"
-        }
-      ]
-    }
-  }
-}
