+++ src/test_data/schema.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
   "$id": "https://ons.gov.uk/bres_schema_for_data.schema.json",
-  "survey_id": "xxx",
+  "survey_id": "xyz",
   "title": "BRES Schema for BRES data",
   "description": "BRES data validation for use with SDS",
   "schemaVersion": "1",
